renamed files to match component names
